requirements.txt
flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.0
requests==2.32.3

# Google Drive API
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0

# AI Services
openai==1.3.7
anthropic==0.7.7

# Document processing (optional)
PyPDF2==3.0.1
python-docx==0.8.11

# Additional utilities
click==8.1.7
gunicorn==21.2.0
